requirements.txt
requests
gradio
pypdf
openai
openai-agents
langchain-community
langchain-openai
faiss-cpu
aiofiles
python-dotenv
asyncio